Fix the broken flake8 build
flake8 part of the build was failing. The problem was that the
`flake8-import-order` plugin was constrained to a very old version
and `flake8` wasn't constrained accordingly, so an old plugin was
using a no-longer-current api and failing.

The reason why the plugin was constrained was to provent a change in the
`google` style that was introduced in v0.6 and conflicted with our
own style. Since v0.7 there is a new `smarkets` style, so the proper
solution is to update to start using that instead.