Add functions to query available prices, ie ticks
The current api in `smarkets.odds` doesn't provide a convenient and
robust way to retrieve all "possible prices", that is, those that
can be used to place bets (usually called "ticks").

This commit fixes that situation by adding iterators that will list
ticks (in probability space) above/below given prices, determine if
if a concrete price is a tick and also define constants for the
minimum and maximum possible prices.