Fix flake8 and pin versions
Build failed after 5b6fdca, because newly released
version 3.2.0 of flake8 was used, which detects
more E305 in our code. Before we used flake8 3.0.4 so these
E305 were not detected (but still existed in our code).

This commit fixes the newly discovered E305 instances
and also pins version of flake8 and its extension
flake8-blind-except, so that in the future we don't
have a build failure because of flake8 unrelated
changes (like in 5b6fdca).