No rounding in Odds.decimal()
Rounding done in decimal() causes incorrect behaviour of snap_to_decimal:

>>> Odds.snap_to_decimal('sell', Decimal('45.045'))

Odds(percent=Decimal('45.05'))

Should return 44.64 instead.

This is because snap_to_decimal_move_out_by first converts to decimal,
so 45.045% gets converted (and rounded) to 2.22, but it should be
2.220002220002220002220002220. Exact 2.22 is exact 45.05% tick, but
45.045% is a bit less, so it should fall to lower tick, 44.64% (worse
price when selling).

The fix is to remove the call to Decimal's quantize(), no rounding.

I don't know why it was done before. My guess is for printing purposes,
so that it prints nicely 2.22 instead of 2.220002220002220002220002220.

If that was the purpose, it shouldn't be. Calculations should be
precise and if one wants to print, he/she uses separate formatting
functions that take precise inner representation and format it to
something nice to the eye (where one can safely round).

Added 2 tests, with previous implementation
test_snap_to_decimal_sell_rounding (which is an example given in
the beggining) was failing.