Add raw bytes to messages sent to global handlers
At the moment, a user of the api cannot save the raw bytes that were received
by the client. This means that one cannot implement reliable logging of what
was received (e.g., for forensics). So now we send a namedtuple that container
both the decoded protobuf message and the raw bytes of the message.